# server/server/src/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"ID": {
				"type": "string"
			},
			"IGroup": {
				"properties": {
					"id": {
						"type": "string"
					},
					"grade": {
						"type": "number",
						"format": "double"
					},
					"subGroup": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"students": {
						"items": {
							"$ref": "#/components/schemas/ID"
						},
						"type": "array"
					},
					"teachers": {
						"items": {
							"$ref": "#/components/schemas/ID"
						},
						"type": "array"
					}
				},
				"required": [
					"grade",
					"subGroup",
					"name",
					"students",
					"teachers"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IStudent": {
				"properties": {
					"id": {
						"type": "string"
					},
					"documentId": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"firstSurname": {
						"type": "string"
					},
					"secondSurname": {
						"type": "string"
					},
					"birthDate": {
						"type": "string",
						"format": "date-time"
					},
					"group": {
						"$ref": "#/components/schemas/ID"
					}
				},
				"required": [
					"documentId",
					"name",
					"firstSurname",
					"birthDate",
					"group"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ITeacher": {
				"properties": {
					"id": {
						"type": "string"
					},
					"documentId": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"firstSurname": {
						"type": "string"
					},
					"secondSurname": {
						"type": "string"
					},
					"groups": {
						"items": {
							"$ref": "#/components/schemas/ID"
						},
						"type": "array"
					}
				},
				"required": [
					"documentId",
					"name",
					"firstSurname",
					"groups"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "@school-portal/server",
		"version": "0.0.0",
		"description": "> TODO: description",
		"license": {
			"name": "MIT"
		},
		"contact": {
			"name": "Alex Vizcaino ",
			"email": "avizcainoluna@gmail.com"
		}
	},
	"openapi": "3.0.0",
	"paths": {
		"/school/group/all": {
			"get": {
				"operationId": "GetGroups",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/IGroup"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"school"
				],
				"security": [],
				"parameters": []
			}
		},
		"/school/group/{id}": {
			"get": {
				"operationId": "GetGroup",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IGroup"
								}
							}
						}
					}
				},
				"tags": [
					"school"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"delete": {
				"operationId": "DeleteGroup",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "boolean"
								}
							}
						}
					}
				},
				"tags": [
					"school"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/school/group": {
			"post": {
				"operationId": "AddGroup",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"tags": [
					"school"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IGroup"
							}
						}
					}
				}
			}
		},
		"/students/all": {
			"get": {
				"operationId": "GetStudents",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/IStudent"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"students"
				],
				"security": [],
				"parameters": []
			}
		},
		"/students/{id}": {
			"get": {
				"operationId": "GetStudent",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IStudent"
								}
							}
						}
					}
				},
				"tags": [
					"students"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"delete": {
				"operationId": "DeleteStudent",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "boolean"
								}
							}
						}
					}
				},
				"tags": [
					"students"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/students": {
			"post": {
				"operationId": "RegisterStudent",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"tags": [
					"students"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IStudent"
							}
						}
					}
				}
			}
		},
		"/teachers/all": {
			"get": {
				"operationId": "GetTeachers",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/ITeacher"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"teachers"
				],
				"security": [],
				"parameters": []
			}
		},
		"/teachers/{id}": {
			"get": {
				"operationId": "GetTeacher",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ITeacher"
								}
							}
						}
					}
				},
				"tags": [
					"teachers"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"delete": {
				"operationId": "DeleteTeacher",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "boolean"
								}
							}
						}
					}
				},
				"tags": [
					"teachers"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateTeacher",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ITeacher"
								}
							}
						}
					}
				},
				"tags": [
					"teachers"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ITeacher"
							}
						}
					}
				}
			}
		},
		"/teachers": {
			"post": {
				"operationId": "RegisterTeacher",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"tags": [
					"teachers"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ITeacher"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}